+++ packages/ui/src/List/components/Content/Caption.tsx
@@ -1,13 +1,25 @@
 import React from 'react';
+import cx from 'classnames';
+
 import {Text} from '../../../Typography';
 
 import styles from './Content.module.scss';
 
-export const Caption: React.FC = (props) => {
-  const {children} = props;
+import type {TextProps} from '../../../Typography';
 
+export const Caption: React.FC<TextProps> = ({
+  children,
+  className,
+  ...props
+}) => {
   return (
-    <Text as="div" size="s" color="muted" className={styles['caption']}>
+    <Text
+      as="div"
+      size="s"
+      color="muted"
+      className={cx(styles['caption'], className)}
+      {...props}
+    >
       {children}
     </Text>
   );